remove the page tsx files so that urls are lowercase with "-" instead of camelcase
