+++ pages/welfare/gail-miller.tsx
@@ -1,16 +1,14 @@
 import Head from "next/head"
-import Layout from "../components/layouts/Layout"
-import PageHeader from "../components/elements/headers/PageHeader"
-import PrimaryButton from "../components/elements/buttons/PrimaryButton"
-import Icon from "../components/elements/icons/Icon"
-import HeroCard from "../components/modules/cards/HeroCard"
-import EventCard, { IEventCard } from "../components/modules/cards/EventCard"
-import ImageCard, { IImageCard } from "../components/modules/cards/ImageCard"
-import BlockQuote from "../components/elements/headers/BlockQuote"
-import { dataInfoCards, dataHeroCard, dataBlockQuote, dataImageCards } from "../data/dataGailMiller"
-import SectionHeader from "../components/elements/headers/SectionHeader"
-import Divider from "../components/elements/dividers/Divider"
-import { isSameOrAfterToday } from "../shared/utils/date.util"
+import Layout from "../../components/layouts/Layout"
+import PageHeader from "../../components/elements/headers/PageHeader"
+import PrimaryButton from "../../components/elements/buttons/PrimaryButton"
+import Icon from "../../components/elements/icons/Icon"
+import HeroCard from "../../components/modules/cards/HeroCard"
+import EventCard, { IEventCard } from "../../components/modules/cards/EventCard"
+import BlockQuote from "../../components/elements/headers/BlockQuote"
+import { dataInfoCards, dataHeroCard, dataBlockQuote } from "../../data/welfare/dataGailMiller"
+import SectionHeader from "../../components/elements/headers/SectionHeader"
+import { isSameOrAfterToday } from "../../shared/utils/date.util"
 
 function GailMiller() {
   return (
@@ -25,7 +23,7 @@ function GailMiller() {
       <Layout>
         <PrimaryButton type="link" className="absolute top-2 md:top-5" link={{ url: "/welfare" }}>
           <Icon name="chevron-left" className="h-5 w-5" />
-          <span className="mr-2 text-sm uppercase font-semibold">Back</span>
+          <span className="mr-2 text-sm uppercase font-semibold">Back To Welfare</span>
         </PrimaryButton>
         <SectionHeader title="Next Opportunity at Gail Miller Homeless Resource Center" />
         <div className="pt-16">
@@ -49,7 +47,7 @@ function GailMiller() {
         </div>
         <PrimaryButton type="link" className="mt-20" link={{ url: "/welfare" }}>
           <Icon name="chevron-left" className="h-5 w-5" />
-          <span className="mr-2 text-sm uppercase font-semibold">Back</span>
+          <span className="mr-2 text-sm uppercase font-semibold">Back To Welfare</span>
         </PrimaryButton>
       </Layout>
     </>
